Навёл немножко порядок. Упросит компоуз
