--- SpringBootServiceAndKafka/src/main/java/mts/service/help/broker/configuration/ConsumerConfiguration.java
@@ -1,49 +1,0 @@
-package mts.service.help.broker.configuration;
-
-import mts.service.help.model.CheeringPhrase;
-import org.apache.kafka.clients.consumer.ConsumerConfig;
-import org.apache.kafka.common.IsolationLevel;
-import org.apache.kafka.common.serialization.StringDeserializer;
-import org.springframework.beans.factory.annotation.Value;
-import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.Configuration;
-import org.springframework.kafka.annotation.EnableKafka;
-import org.springframework.kafka.config.ConcurrentKafkaListenerContainerFactory;
-import org.springframework.kafka.core.ConsumerFactory;
-import org.springframework.kafka.core.DefaultKafkaConsumerFactory;
-import org.springframework.kafka.support.serializer.JsonDeserializer;
-
-import java.util.HashMap;
-import java.util.Map;
-
-
-@EnableKafka
-@Configuration
-public class ConsumerConfiguration {
-
-    @Value("${kafka.server}")
-    private String kafkaServer;
-
-    @Value("${kafka.group.id}")
-    private String kafkaGroupId;
-
-    public ConsumerFactory<String, CheeringPhrase> cheeringConsumerFactory() {
-        Map<String, Object> props = new HashMap<>();
-        props.put(ConsumerConfig.BOOTSTRAP_SERVERS_CONFIG, kafkaServer);
-        props.put(ConsumerConfig.GROUP_ID_CONFIG, kafkaGroupId);
-        props.put(ConsumerConfig.AUTO_OFFSET_RESET_CONFIG, "latest");
-//        props.put(ConsumerConfig.ISOLATION_LEVEL_CONFIG, IsolationLevel.READ_COMMITTED);
-
-        return new DefaultKafkaConsumerFactory<>(props, new StringDeserializer(), new JsonDeserializer<>(CheeringPhrase.class));
-    }
-
-    @Bean
-    public ConcurrentKafkaListenerContainerFactory<String, CheeringPhrase> cheeringKafkaListenerContainerFactory() {
-        ConcurrentKafkaListenerContainerFactory<String, CheeringPhrase> factory = new ConcurrentKafkaListenerContainerFactory<>();
-        factory.setConsumerFactory(cheeringConsumerFactory());
-
-        return factory;
-    }
-
-
-}
