--- SpringBootServiceAndKafka/src/main/java/mts/service/help/broker/configuration/ProducerConfiguration.java
@@ -1,43 +1,0 @@
-package mts.service.help.broker.configuration;
-
-import mts.service.help.model.CheeringPhrase;
-import org.apache.kafka.clients.producer.ProducerConfig;
-import org.apache.kafka.common.serialization.StringSerializer;
-import org.springframework.beans.factory.annotation.Value;
-import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.Configuration;
-import org.springframework.kafka.core.DefaultKafkaProducerFactory;
-import org.springframework.kafka.core.KafkaTemplate;
-import org.springframework.kafka.core.ProducerFactory;
-import org.springframework.kafka.support.serializer.JsonSerializer;
-
-import java.util.HashMap;
-import java.util.Map;
-
-@Configuration
-public class ProducerConfiguration {
-
-    @Value("${kafka.server}")
-    private String kafkaServer;
-
-    @Value("producerServer")
-    private String kafkaProducerId;
-
-    @Bean
-    public ProducerFactory<String, CheeringPhrase> cheeringProducerFactory() {
-        Map<String, Object> configProps = new HashMap<>();
-        configProps.put(ProducerConfig.BOOTSTRAP_SERVERS_CONFIG, kafkaServer);
-        configProps.put(ProducerConfig.KEY_SERIALIZER_CLASS_CONFIG, StringSerializer.class);
-        configProps.put(ProducerConfig.VALUE_SERIALIZER_CLASS_CONFIG, JsonSerializer.class);
-        configProps.put(ProducerConfig.CLIENT_ID_CONFIG, kafkaProducerId);
-        configProps.put(ProducerConfig.ENABLE_IDEMPOTENCE_CONFIG, "true");
-
-        return new DefaultKafkaProducerFactory<>(configProps);
-    }
-
-    @Bean
-    public KafkaTemplate<String, CheeringPhrase> cheeringTemplate() {
-        return new KafkaTemplate<>(cheeringProducerFactory());
-    }
-
-}
