+++ SpringBootServiceAndKafka/src/main/java/mts/service/help/broker/interfaces/Producer.java
@@ -1,7 +1,5 @@
 package mts.service.help.broker.interfaces;
 
-import mts.service.help.model.CheeringPhrase;
-
 public interface Producer<T> {
 
     void sendMessage(T t);
