+++ SpringBootServiceAndKafka/src/main/java/mts/service/help/config/AppConfig.java
@@ -4,7 +4,6 @@ import mts.service.help.repository.CheeringInMemRepository;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
 
-
 @Configuration
 public class AppConfig {
