+++ SpringBootServiceAndKafka/src/test/java/mts/service/help/services/CheeringServiceImpTest.java
@@ -1,6 +1,5 @@
 package mts.service.help.services;
 
-import com.fasterxml.jackson.databind.ObjectMapper;
 import mts.service.help.repository.CheeringInMemRepository;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
@@ -8,8 +7,6 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.autoconfigure.web.servlet.AutoConfigureMockMvc;
 import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.test.web.servlet.MockMvc;
-
-import static org.junit.jupiter.api.Assertions.*;
 
 @SpringBootTest
 @AutoConfigureMockMvc
